Latest Code with all detailed Introduction/Data Preprocessing/ Explanations/ Result Interpretations and Recommendations.
This is the Final project file with Results Interpretation and recommendation. We dont have the best accuracy as the MSE all are very high. I modified the Naradya Watson - removed redundant code and added MAE for both Lasso and MSE and added Scaled data for Naradya watson. Lets discuss what to eliminate and simplify and fine tune the project.